+++ src/XFM2_params.h
@@ -1,3 +1,6 @@
+/* -*- C++ -*- */
+#pragma once
+
 // This file should really contain the definitions for each parameter
 // Define other values (globals)
 // #define MIDI_Channel_0              * 10 // 0 (omni) or 1 - 16
@@ -324,327 +327,3 @@
 #define PRM_ARP_TEMPO_FINE          452
 #define PRM_ARP_DIVISION            453
 #define PRM_ARP_OCTAVES             454
-
-// Define integers for all of the internal Paramters
-
-int ALGO0 = 0;
-int ALGO1 = 0;
-int ALGO2 = 0;
-int ALGO3 = 0;
-int ALGO4 = 0;
-int ALGO5 = 0;
-int FEEDBACK0 = 0;
-int FEEDBACK1 = 0;
-int FEEDBACK2 = 0;
-int FEEDBACK3 = 0;
-int FEEDBACK4 = 0;
-int FEEDBACK5 = 0;
-int OSC_SYNC = 0;
-int OSC_MODE = 0;
-int RATIO0 = 0;
-int RATIO1 = 0;
-int RATIO2 = 0;
-int RATIO3 = 0;
-int RATIO4 = 0;
-int RATIO5 = 0;
-int RATIO_FINE0 = 0;
-int RATIO_FINE1 = 0;
-int RATIO_FINE2 = 0;
-int RATIO_FINE3 = 0;
-int RATIO_FINE4 = 0;
-int RATIO_FINE5 = 0;
-int FINE0 = 0;
-int FINE1 = 0;
-int FINE2 = 0;
-int FINE3 = 0;
-int FINE4 = 0;
-int FINE5 = 0;
-int LEVEL0 = 0;
-int LEVEL1 = 0;
-int LEVEL2 = 0;
-int LEVEL3 = 0;
-int LEVEL4 = 0;
-int LEVEL5 = 0;
-int VEL_SENS0 = 0;
-int VEL_SENS1 = 0;
-int VEL_SENS2 = 0;
-int VEL_SENS3 = 0;
-int VEL_SENS4 = 0;
-int VEL_SENS5 = 0;
-int KEY_BP0 = 0;
-int KEY_BP1 = 0;
-int KEY_BP2 = 0;
-int KEY_BP3 = 0;
-int KEY_BP4 = 0;
-int KEY_BP5 = 0;
-int KEY_LDEPTH0 = 0;
-int KEY_LDEPTH1 = 0;
-int KEY_LDEPTH2 = 0;
-int KEY_LDEPTH3 = 0;
-int KEY_LDEPTH4 = 0;
-int KEY_LDEPTH5 = 0;
-int KEY_RDEPTH0 = 0;
-int KEY_RDEPTH1 = 0;
-int KEY_RDEPTH2 = 0;
-int KEY_RDEPTH3 = 0;
-int KEY_RDEPTH4 = 0;
-int KEY_RDEPTH5 = 0;
-int KEY_LCURVE0 = 0;
-int KEY_LCURVE1 = 0;
-int KEY_LCURVE2 = 0;
-int KEY_LCURVE3 = 0;
-int KEY_LCURVE4 = 0;
-int KEY_LCURVE5 = 0;
-int KEY_RCURVE0 = 0;
-int KEY_RCURVE1 = 0;
-int KEY_RCURVE2 = 0;
-int KEY_RCURVE3 = 0;
-int KEY_RCURVE4 = 0;
-int KEY_RCURVE5 = 0;
-int L1_0 = 0;
-int L1_1 = 0;
-int L1_2 = 0;
-int L1_3 = 0;
-int L1_4 = 0;
-int L1_5 = 0;
-int L2_0 = 0;
-int L2_1 = 0;
-int L2_2 = 0;
-int L2_3 = 0;
-int L2_4 = 0;
-int L2_5 = 0;
-int L3_0 = 0;
-int L3_1 = 0;
-int L3_2 = 0;
-int L3_3 = 0;
-int L3_4 = 0;
-int L3_5 = 0;
-int L4_0 = 0;
-int L4_1 = 0;
-int L4_2 = 0;
-int L4_3 = 0;
-int L4_4 = 0;
-int L4_5 = 0;
-int R1_0 = 0;
-int R1_1 = 0;
-int R1_2 = 0;
-int R1_3 = 0;
-int R1_4 = 0;
-int R1_5 = 0;
-int R2_0 = 0;
-int R2_1 = 0;
-int R2_2 = 0;
-int R2_3 = 0;
-int R2_4 = 0;
-int R2_5 = 0;
-int R3_0 = 0;
-int R3_1 = 0;
-int R3_2 = 0;
-int R3_3 = 0;
-int R3_4 = 0;
-int R3_5 = 0;
-int R4_0 = 0;
-int R4_1 = 0;
-int R4_2 = 0;
-int R4_3 = 0;
-int R4_4 = 0;
-int R4_5 = 0;
-int L1_P = 0;
-int L2_P = 0;
-int L3_P = 0;
-int L4_P = 0;
-int R1_P = 0;
-int R2_P = 0;
-int R3_P = 0;
-int R4_P = 0;
-int EG_PITCH_RANGE = 0;
-int EG_PITCH_RANGE_VELO = 0;
-int RATE_KEY0 = 0;
-int RATE_KEY1 = 0;
-int RATE_KEY2 = 0;
-int RATE_KEY3 = 0;
-int RATE_KEY4 = 0;
-int RATE_KEY5 = 0;
-int RATE_KEYP = 0;
-int LFO_PITCH_DEPTH = 0;
-int LFO_AMP_DEPTH = 0;
-int LFO_SPEED = 0;
-int LFO_SYNC = 0;
-int LFO_WAVE = 0;
-int LFO_FADE = 0;
-int LFO_PITCH_DEPTH_WHEEL = 0;
-int LFO_AMP_DEPTH_WHEEL = 0;
-int LFO_PITCH_DEPTH_AFTER = 0;
-int LFO_AMP_DEPTH_AFTER = 0;
-int LFO_AMS0 = 0;
-int LFO_AMS1 = 0;
-int LFO_AMS2 = 0;
-int LFO_AMS3 = 0;
-int LFO_AMS4 = 0;
-int LFO_AMS5 = 0;
-int BEND_RANGE_UP = 0;
-int BEND_RANGE_DN = 0;
-int TRANSPOSE = 0;
-int VOLUME = 0;
-int L0_0 = 0;
-int L0_1 = 0;
-int L0_2 = 0;
-int L0_3 = 0;
-int L0_4 = 0;
-int L0_5 = 0;
-int R0_0 = 0;
-int R0_1 = 0;
-int R0_2 = 0;
-int R0_3 = 0;
-int R0_4 = 0;
-int R0_5 = 0;
-int L5_0 = 0;
-int L5_1 = 0;
-int L5_2 = 0;
-int L5_3 = 0;
-int L5_4 = 0;
-int L5_5 = 0;
-int R5_0 = 0;
-int R5_1 = 0;
-int R5_2 = 0;
-int R5_3 = 0;
-int R5_4 = 0;
-int R5_5 = 0;
-int L0_P = 0;
-int R0_P = 0;
-int L5_P = 0;
-int R5_P = 0;
-int LFO_PITCH_DEPTH_BREATH = 0;
-int LFO_AMP_DEPTH_BREATH = 0;
-int LFO_PITCH_DEPTH_FOOT = 0;
-int LFO_AMP_DEPTH_FOOT = 0;
-int VOLUME_DEPTH_AFTER = 0;
-int VOLUME_DEPTH_WHEEL = 0;
-int VOLUME_DEPTH_BREATH = 0;
-int VOLUME_DEPTH_FOOT = 0;
-int PITCH_DEPTH_AFTER = 0;
-int PITCH_DEPTH_BREATH = 0;
-int PITCH_DEPTH_FOOT = 0;
-int PITCH_DEPTH_RND = 0;
-int PAN = 0;
-int LFO_PMS0 = 0;
-int LFO_PMS1 = 0;
-int LFO_PMS2 = 0;
-int LFO_PMS3 = 0;
-int LFO_PMS4 = 0;
-int LFO_PMS5 = 0;
-int LEGATO = 0;
-int PORTA_MODE = 0;
-int PORTA_TIME = 0;
-int WAVEFORM0 = 0;
-int WAVEFORM1 = 0;
-int WAVEFORM2 = 0;
-int WAVEFORM3 = 0;
-int WAVEFORM4 = 0;
-int WAVEFORM5 = 0;
-int VELO_OFFSET = 0;
-int VELO_CURVE = 0;
-int EG_LOOP = 0;
-int EG_LOOP_SEG = 0;
-int EG_RESTART = 0;
-int TUNING = 0;
-int LEVEL0_L = 0;
-int LEVEL0_R = 0;
-int LEVEL1_L = 0;
-int LEVEL1_R = 0;
-int LEVEL2_L = 0;
-int LEVEL2_R = 0;
-int LEVEL3_L = 0;
-int LEVEL3_R = 0;
-int LEVEL4_L = 0;
-int LEVEL4_R = 0;
-int LEVEL5_L = 0;
-int LEVEL5_R = 0;
-int WAVEFORM2_0 = 0;
-int WAVEFORM2_1 = 0;
-int WAVEFORM2_2 = 0;
-int WAVEFORM2_3 = 0;
-int WAVEFORM2_4 = 0;
-int WAVEFORM2_5 = 0;
-int WMODE0 = 0;
-int WMODE1 = 0;
-int WMODE2 = 0;
-int WMODE3 = 0;
-int WMODE4 = 0;
-int WMODE5 = 0;
-int WRATIO0 = 0;
-int WRATIO1 = 0;
-int WRATIO2 = 0;
-int WRATIO3 = 0;
-int WRATIO4 = 0;
-int WRATIO5 = 0;
-int OSC_PHASE0 = 0;
-int OSC_PHASE1 = 0;
-int OSC_PHASE2 = 0;
-int OSC_PHASE3 = 0;
-int OSC_PHASE4 = 0;
-int OSC_PHASE5 = 0;
-int SDLY_DRY = 0;
-int SDLY_WET = 0;
-int SDLY_MODE = 0;
-int SDLY_TIME = 0;
-int SDLY_FEEDBACK = 0;
-int SDLY_FEEDLO = 0;
-int SDLY_FEEDHI = 0;
-int SDLY_TEMPO = 0;
-int SDLY_NUM = 0;
-int SDLY_DEN = 0;
-int PHASER_DRY = 0;
-int PHASER_WET = 0;
-int PHASER_MODE = 0;
-int PHASER_DEPTH = 0;
-int PHASER_SPEED = 0;
-int PHASER_FEEDBACK = 0;
-int PHASER_OFFSET = 0;
-int PHASER_STAGES = 0;
-int PHASER_PHASE = 0;
-int PHASER_WAVEFORM = 0;
-int FILTER_LO = 0;
-int FILTER_HI = 0;
-int AM_SPEED = 0;
-int AM_RANGE = 0;
-int AM_DEPTH = 0;
-int AM_PHASE = 0;
-int FIR_IMPULSE = 0;
-int WAVESHAPER_DRY = 0;
-int WAVESHAPER_WET = 0;
-int WAVESHAPER_GAIN_PRE = 0;
-int WAVESHAPER_GAIN_POST = 0;
-int WAVESHAPER_SHAPE = 0;
-int CHORUS_DRY = 0;
-int CHORUS_WET = 0;
-int CHORUS_MODE = 0;
-int CHORUS_SPEED = 0;
-int CHORUS_DEPTH = 0;
-int CHORUS_FEEDBACK = 0;
-int CHORUS_PHASE = 0;
-int DECIMATOR_DEPTH = 0;
-int BITCRUSHER_DEPTH = 0;
-int REVERB_DRY = 0;
-int REVERB_WET = 0;
-int REVERB_MODE = 0;
-int REVERB_DECAY = 0;
-int REVERB_DAMP = 0;
-int REVERB_D0 = 0;
-int REVERB_D1 = 0;
-int REVERB_D2 = 0;
-int FX_ROUTING = 0;
-int OUTPUT_LEVEL = 0;
-int CONTROL_1H = 0;
-int CONTROL_1L = 0;
-int CONTROL_2H = 0;
-int CONTROL_2L = 0;
-int CONTROL_3H = 0;
-int CONTROL_3L = 0;
-int CONTROL_4H = 0;
-int CONTROL_4L = 0;
-int ARP_MODE = 0;
-int ARP_TEMPO = 0;
-int ARP_TEMPO_FINE = 0;
-int ARP_DIVISION = 0;
-int ARP_OCTAVES = 0;
