+++ src/XFM2_vars.h
@@ -1,0 +1,328 @@
+/* -*- C++ -*- */
+#pragma once
+
+#include "XFM2_params.h"
+
+// Define integers for all of the internal Paramters
+
+extern int ALGO0;
+extern int ALGO1;
+extern int ALGO2;
+extern int ALGO3;
+extern int ALGO4;
+extern int ALGO5;
+extern int FEEDBACK0;
+extern int FEEDBACK1;
+extern int FEEDBACK2;
+extern int FEEDBACK3;
+extern int FEEDBACK4;
+extern int FEEDBACK5;
+extern int OSC_SYNC;
+extern int OSC_MODE;
+extern int RATIO0;
+extern int RATIO1;
+extern int RATIO2;
+extern int RATIO3;
+extern int RATIO4;
+extern int RATIO5;
+extern int RATIO_FINE0;
+extern int RATIO_FINE1;
+extern int RATIO_FINE2;
+extern int RATIO_FINE3;
+extern int RATIO_FINE4;
+extern int RATIO_FINE5;
+extern int FINE0;
+extern int FINE1;
+extern int FINE2;
+extern int FINE3;
+extern int FINE4;
+extern int FINE5;
+extern int LEVEL0;
+extern int LEVEL1;
+extern int LEVEL2;
+extern int LEVEL3;
+extern int LEVEL4;
+extern int LEVEL5;
+extern int VEL_SENS0;
+extern int VEL_SENS1;
+extern int VEL_SENS2;
+extern int VEL_SENS3;
+extern int VEL_SENS4;
+extern int VEL_SENS5;
+extern int KEY_BP0;
+extern int KEY_BP1;
+extern int KEY_BP2;
+extern int KEY_BP3;
+extern int KEY_BP4;
+extern int KEY_BP5;
+extern int KEY_LDEPTH0;
+extern int KEY_LDEPTH1;
+extern int KEY_LDEPTH2;
+extern int KEY_LDEPTH3;
+extern int KEY_LDEPTH4;
+extern int KEY_LDEPTH5;
+extern int KEY_RDEPTH0;
+extern int KEY_RDEPTH1;
+extern int KEY_RDEPTH2;
+extern int KEY_RDEPTH3;
+extern int KEY_RDEPTH4;
+extern int KEY_RDEPTH5;
+extern int KEY_LCURVE0;
+extern int KEY_LCURVE1;
+extern int KEY_LCURVE2;
+extern int KEY_LCURVE3;
+extern int KEY_LCURVE4;
+extern int KEY_LCURVE5;
+extern int KEY_RCURVE0;
+extern int KEY_RCURVE1;
+extern int KEY_RCURVE2;
+extern int KEY_RCURVE3;
+extern int KEY_RCURVE4;
+extern int KEY_RCURVE5;
+extern int L1_0;
+extern int L1_1;
+extern int L1_2;
+extern int L1_3;
+extern int L1_4;
+extern int L1_5;
+extern int L2_0;
+extern int L2_1;
+extern int L2_2;
+extern int L2_3;
+extern int L2_4;
+extern int L2_5;
+extern int L3_0;
+extern int L3_1;
+extern int L3_2;
+extern int L3_3;
+extern int L3_4;
+extern int L3_5;
+extern int L4_0;
+extern int L4_1;
+extern int L4_2;
+extern int L4_3;
+extern int L4_4;
+extern int L4_5;
+extern int R1_0;
+extern int R1_1;
+extern int R1_2;
+extern int R1_3;
+extern int R1_4;
+extern int R1_5;
+extern int R2_0;
+extern int R2_1;
+extern int R2_2;
+extern int R2_3;
+extern int R2_4;
+extern int R2_5;
+extern int R3_0;
+extern int R3_1;
+extern int R3_2;
+extern int R3_3;
+extern int R3_4;
+extern int R3_5;
+extern int R4_0;
+extern int R4_1;
+extern int R4_2;
+extern int R4_3;
+extern int R4_4;
+extern int R4_5;
+extern int L1_P;
+extern int L2_P;
+extern int L3_P;
+extern int L4_P;
+extern int R1_P;
+extern int R2_P;
+extern int R3_P;
+extern int R4_P;
+extern int EG_PITCH_RANGE;
+extern int EG_PITCH_RANGE_VELO;
+extern int RATE_KEY0;
+extern int RATE_KEY1;
+extern int RATE_KEY2;
+extern int RATE_KEY3;
+extern int RATE_KEY4;
+extern int RATE_KEY5;
+extern int RATE_KEYP;
+extern int LFO_PITCH_DEPTH;
+extern int LFO_AMP_DEPTH;
+extern int LFO_SPEED;
+extern int LFO_SYNC;
+extern int LFO_WAVE;
+extern int LFO_FADE;
+extern int LFO_PITCH_DEPTH_WHEEL;
+extern int LFO_AMP_DEPTH_WHEEL;
+extern int LFO_PITCH_DEPTH_AFTER;
+extern int LFO_AMP_DEPTH_AFTER;
+extern int LFO_AMS0;
+extern int LFO_AMS1;
+extern int LFO_AMS2;
+extern int LFO_AMS3;
+extern int LFO_AMS4;
+extern int LFO_AMS5;
+extern int BEND_RANGE_UP;
+extern int BEND_RANGE_DN;
+extern int TRANSPOSE;
+extern int VOLUME;
+extern int L0_0;
+extern int L0_1;
+extern int L0_2;
+extern int L0_3;
+extern int L0_4;
+extern int L0_5;
+extern int R0_0;
+extern int R0_1;
+extern int R0_2;
+extern int R0_3;
+extern int R0_4;
+extern int R0_5;
+extern int L5_0;
+extern int L5_1;
+extern int L5_2;
+extern int L5_3;
+extern int L5_4;
+extern int L5_5;
+extern int R5_0;
+extern int R5_1;
+extern int R5_2;
+extern int R5_3;
+extern int R5_4;
+extern int R5_5;
+extern int L0_P;
+extern int R0_P;
+extern int L5_P;
+extern int R5_P;
+extern int LFO_PITCH_DEPTH_BREATH;
+extern int LFO_AMP_DEPTH_BREATH;
+extern int LFO_PITCH_DEPTH_FOOT;
+extern int LFO_AMP_DEPTH_FOOT;
+extern int VOLUME_DEPTH_AFTER;
+extern int VOLUME_DEPTH_WHEEL;
+extern int VOLUME_DEPTH_BREATH;
+extern int VOLUME_DEPTH_FOOT;
+extern int PITCH_DEPTH_AFTER;
+extern int PITCH_DEPTH_BREATH;
+extern int PITCH_DEPTH_FOOT;
+extern int PITCH_DEPTH_RND;
+extern int PAN;
+extern int LFO_PMS0;
+extern int LFO_PMS1;
+extern int LFO_PMS2;
+extern int LFO_PMS3;
+extern int LFO_PMS4;
+extern int LFO_PMS5;
+extern int LEGATO;
+extern int PORTA_MODE;
+extern int PORTA_TIME;
+extern int WAVEFORM0;
+extern int WAVEFORM1;
+extern int WAVEFORM2;
+extern int WAVEFORM3;
+extern int WAVEFORM4;
+extern int WAVEFORM5;
+extern int VELO_OFFSET;
+extern int VELO_CURVE;
+extern int EG_LOOP;
+extern int EG_LOOP_SEG;
+extern int EG_RESTART;
+extern int TUNING;
+extern int LEVEL0_L;
+extern int LEVEL0_R;
+extern int LEVEL1_L;
+extern int LEVEL1_R;
+extern int LEVEL2_L;
+extern int LEVEL2_R;
+extern int LEVEL3_L;
+extern int LEVEL3_R;
+extern int LEVEL4_L;
+extern int LEVEL4_R;
+extern int LEVEL5_L;
+extern int LEVEL5_R;
+extern int WAVEFORM2_0;
+extern int WAVEFORM2_1;
+extern int WAVEFORM2_2;
+extern int WAVEFORM2_3;
+extern int WAVEFORM2_4;
+extern int WAVEFORM2_5;
+extern int WMODE0;
+extern int WMODE1;
+extern int WMODE2;
+extern int WMODE3;
+extern int WMODE4;
+extern int WMODE5;
+extern int WRATIO0;
+extern int WRATIO1;
+extern int WRATIO2;
+extern int WRATIO3;
+extern int WRATIO4;
+extern int WRATIO5;
+extern int OSC_PHASE0;
+extern int OSC_PHASE1;
+extern int OSC_PHASE2;
+extern int OSC_PHASE3;
+extern int OSC_PHASE4;
+extern int OSC_PHASE5;
+extern int SDLY_DRY;
+extern int SDLY_WET;
+extern int SDLY_MODE;
+extern int SDLY_TIME;
+extern int SDLY_FEEDBACK;
+extern int SDLY_FEEDLO;
+extern int SDLY_FEEDHI;
+extern int SDLY_TEMPO;
+extern int SDLY_NUM;
+extern int SDLY_DEN;
+extern int PHASER_DRY;
+extern int PHASER_WET;
+extern int PHASER_MODE;
+extern int PHASER_DEPTH;
+extern int PHASER_SPEED;
+extern int PHASER_FEEDBACK;
+extern int PHASER_OFFSET;
+extern int PHASER_STAGES;
+extern int PHASER_PHASE;
+extern int PHASER_WAVEFORM;
+extern int FILTER_LO;
+extern int FILTER_HI;
+extern int AM_SPEED;
+extern int AM_RANGE;
+extern int AM_DEPTH;
+extern int AM_PHASE;
+extern int FIR_IMPULSE;
+extern int WAVESHAPER_DRY;
+extern int WAVESHAPER_WET;
+extern int WAVESHAPER_GAIN_PRE;
+extern int WAVESHAPER_GAIN_POST;
+extern int WAVESHAPER_SHAPE;
+extern int CHORUS_DRY;
+extern int CHORUS_WET;
+extern int CHORUS_MODE;
+extern int CHORUS_SPEED;
+extern int CHORUS_DEPTH;
+extern int CHORUS_FEEDBACK;
+extern int CHORUS_PHASE;
+extern int DECIMATOR_DEPTH;
+extern int BITCRUSHER_DEPTH;
+extern int REVERB_DRY;
+extern int REVERB_WET;
+extern int REVERB_MODE;
+extern int REVERB_DECAY;
+extern int REVERB_DAMP;
+extern int REVERB_D0;
+extern int REVERB_D1;
+extern int REVERB_D2;
+extern int FX_ROUTING;
+extern int OUTPUT_LEVEL;
+extern int CONTROL_1H;
+extern int CONTROL_1L;
+extern int CONTROL_2H;
+extern int CONTROL_2L;
+extern int CONTROL_3H;
+extern int CONTROL_3L;
+extern int CONTROL_4H;
+extern int CONTROL_4L;
+extern int ARP_MODE;
+extern int ARP_TEMPO;
+extern int ARP_TEMPO_FINE;
+extern int ARP_DIVISION;
+extern int ARP_OCTAVES;
