+++ src/base.cpp
@@ -1,0 +1,3 @@
+#include "base.h"
+
+U8G2_SSD1306_128X64_NONAME_F_HW_I2C u8g2(U8G2_R0, U8X8_PIN_NONE);
