+++ src/base.h
@@ -1,0 +1,11 @@
+/* -*- C++ -*- */
+#pragma once
+
+//XFM2 parameter setting - From Rene's code
+#include "xfm2.h"
+
+//graphics lib
+#include <U8g2lib.h>
+
+//init base hardware
+extern U8G2_SSD1306_128X64_NONAME_F_HW_I2C u8g2;
