+++ src/xfm2.cpp
@@ -1,0 +1,39 @@
+#include <Arduino.h>
+#include "xfm2.h"
+
+void set_unit( int unit ) {
+    // Default to unit 1
+    Serial1.write( unit == 2 ? '2' : '1' );
+}
+
+// Code for setting the parameter on the XFM2
+void set_parameter( int param, int value ) {
+    Serial1.write( 's' ); // 's' = Set Parameter
+
+    if( param > 255 ) {
+        // Parameters above 255 have a two-byte format: b1 = 255, b2 = x-256
+        Serial1.write( 255 );
+        Serial1.write( param - 256  );
+        Serial1.write( value );
+    }
+    else {
+        Serial1.write( param );
+        Serial1.write( value );
+    }
+}
+
+// Code for getting a selected parameter on the XFM2
+void get_parameter( int param, int value ) {
+    Serial1.write( 'g' ); // 'g' = Get Parameter
+
+    if( param > 255 ) {
+        // Parameters above 255 have a two-byte format: b1 = 255, b2 = x-256
+        Serial1.write( 255 );
+        Serial1.write( param - 256  );
+        Serial1.write( value );
+    }
+    else {
+        Serial1.write( param );
+        Serial1.write( value );
+    }
+}
