+++ src/xfm2.h
@@ -1,0 +1,11 @@
+/* -*- C++ -*- */
+#pragma once
+
+//XFM2 parameter setting - From Rene's code
+
+void set_parameter( int param, int value );
+void set_unit( int unit );
+// Code for setting the parameter on the XFM2
+void set_parameter( int param, int value );
+// Code for getting a selected parameter on the XFM2
+void get_parameter( int param, int value );
